Add widget_helper.c
