+++ include/LCUI/gui/Makefile.am
@@ -3,5 +3,5 @@ SUBDIRS=widget
 # Headers to install
 pkginclude_HEADERS = widget_base.h widget_task.h widget_prototype.h \
 widget_style.h widget_event.h widget_paint.h widget.h css_library.h \
-css_parser.h css_fontstyle.h builder.h metrics.h
+widget_helper.h css_parser.h css_fontstyle.h builder.h metrics.h
 pkgincludedir=$(prefix)/include/LCUI/gui
