+++ src/gui/widget_base.c
@@ -1239,74 +1239,6 @@ void Widget_UpdateProps( LCUI_Widget w )
 	}
 }
 
-void Widget_SetPadding( LCUI_Widget w, float top, float right, 
-			float bottom, float left )
-{
-	Widget_SetStyle( w, key_padding_top, top, px );
-	Widget_SetStyle( w, key_padding_right, right, px );
-	Widget_SetStyle( w, key_padding_bottom, bottom, px );
-	Widget_SetStyle( w, key_padding_left, left, px );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_SetMargin( LCUI_Widget w, float top, float right,
-		       float bottom, float left )
-{
-	Widget_SetStyle( w, key_margin_top, top, px );
-	Widget_SetStyle( w, key_margin_right, right, px );
-	Widget_SetStyle( w, key_margin_bottom, bottom, px );
-	Widget_SetStyle( w, key_margin_left, left, px );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_Move( LCUI_Widget w, float left, float top )
-{
-	Widget_SetStyle( w, key_top, top, px );
-	Widget_SetStyle( w, key_left, left, px );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_Resize( LCUI_Widget w, float width, float height )
-{
-	Widget_SetStyle( w, key_width, width, px );
-	Widget_SetStyle( w, key_height, height, px );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_Show( LCUI_Widget w )
-{
-	Widget_SetStyle( w, key_visible, TRUE, int );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_Hide( LCUI_Widget w )
-{
-	Widget_SetStyle( w, key_visible, FALSE, int );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_SetPosition( LCUI_Widget w, LCUI_StyleValue position )
-{
-	Widget_SetStyle( w, key_position, position, style );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_SetBoxSizing( LCUI_Widget w, LCUI_StyleValue sizing )
-{
-	Widget_SetStyle( w, key_box_sizing, sizing, style );
-	Widget_UpdateStyle( w, FALSE );
-}
-
-void Widget_SetDisabled( LCUI_Widget w, LCUI_BOOL disabled )
-{
-	w->disabled = disabled;
-	if( w->disabled ) {
-		Widget_AddStatus( w, "disabled" );
-	} else {
-		Widget_RemoveStatus( w, "disabled" );
-	}
-}
-
 int Widget_SetAttributeEx( LCUI_Widget w, const char *name, void *value,
 			   int value_type, void( *value_destructor )(void*) )
 {
